Revert bag slot index type to uint8_t for now
was causing incompatibility on vanilla servers - need to add net element filters for legacy/modified servers first

+++ source/game/StarInventoryTypes.hpp
@@ -19,7 +19,7 @@ enum class EquipmentSlot : uint8_t {
 };
 extern EnumMap<EquipmentSlot> const EquipmentSlotNames;
 
-typedef pair<String, unsigned> BagSlot;
+typedef pair<String, uint8_t> BagSlot;
 
 strong_typedef(Empty, SwapSlot);
 strong_typedef(Empty, TrashSlot);
